Create Machine learning approach for rainfall prediction
India is an agricultural country and its economy is largely based upon crop productivity and rainfall. For analysing the crop productivity, rainfall prediction is require and necessary to all farmers. Rainfall Prediction is the application of science and technology to predict the state of the atmosphere. It is important to exactly determine the rainfall for effective use of water resources, crop productivity and pre planning of water structures. Using different data mining techniques it can predict rainfall. Data mining techniques are used to estimate the rainfall numerically. This paper focuses some of the popular data mining algorithms for rainfall prediction. CatBoost classifier, Random Forest, K-Nearest Neighbour algorithm, Logistic Regression , Decision Tree are some of the algorithms compared in this paper. From that comparison, it can analyze which method gives better accuracy for rainfall prediction